+++ crates/bitwarden-crypto/src/keys/mod.rs
@@ -1,16 +1,7 @@
-mod key_encryptable;
-pub(crate) use key_encryptable::KeyEncryptableWithContentType;
-pub use key_encryptable::{CryptoKey, KeyContainer, KeyDecryptable, KeyEncryptable};
 mod master_key;
 pub use master_key::{HashPurpose, MasterKey};
 mod shareable_key;
 pub use shareable_key::derive_shareable_key;
-mod symmetric_crypto_key;
-#[cfg(test)]
-pub use symmetric_crypto_key::derive_symmetric_key;
-pub use symmetric_crypto_key::{
-    Aes256CbcHmacKey, Aes256CbcKey, SymmetricCryptoKey, XChaCha20Poly1305Key,
-};
 mod user_key;
 pub use user_key::UserKey;
 mod device_key;
