+++ crates/bitwarden-crypto/src/lib.rs
@@ -12,12 +12,8 @@
 #[global_allocator]
 static ALLOC: ZeroizingAllocator<std::alloc::System> = ZeroizingAllocator(std::alloc::System);
 
-mod aes;
 mod content_format;
 pub use content_format::*;
-mod enc_string;
-pub use enc_string::EncString;
-pub(crate) use enc_string::{from_b64_vec, split_enc_string};
 mod error;
 pub use error::CryptoError;
 pub(crate) use error::Result;
@@ -41,11 +37,10 @@ mod signing;
 pub use signing::*;
 mod public_key_encryption;
 pub use public_key_encryption::*;
+mod symmetric_encryption;
+pub use symmetric_encryption::*;
 mod traits;
-mod xchacha20;
-pub use traits::{
-    CompositeEncryptable, Decryptable, IdentifyKey, KeyId, KeyIds, PrimitiveEncryptable,
-};
+pub use traits::{IdentifyKey, KeyId, KeyIds};
 pub use zeroizing_alloc::ZeroAlloc as ZeroizingAllocator;
 
 #[cfg(feature = "uniffi")]
