+++ crates/bitwarden-crypto/src/symmetric_encryption/decryptable.rs
