+++ crates/bitwarden-crypto/src/symmetric_encryption/hazmat/aes.rs
