+++ crates/bitwarden-crypto/src/symmetric_encryption/key_encryptable.rs
