+++ crates/bitwarden-crypto/src/symmetric_encryption/mod.rs
@@ -1,0 +1,20 @@
+mod cose;
+mod decryptable;
+pub use decryptable::*;
+mod encryptable;
+pub use encryptable::*;
+mod enc_string;
+pub use enc_string::*;
+mod symmetric_crypto_key;
+pub use symmetric_crypto_key::{
+    Aes256CbcHmacKey, Aes256CbcKey, SymmetricCryptoKey, XChaCha20Poly1305Key,
+};
+mod key_encryptable;
+pub(crate) use key_encryptable::KeyEncryptableWithContentType;
+pub use key_encryptable::{CryptoKey, KeyContainer, KeyDecryptable, KeyEncryptable};
+mod hazmat;
+pub(crate) mod key_wrap;
+mod util;
+#[cfg(test)]
+pub use symmetric_crypto_key::derive_symmetric_key;
+pub(crate) use util::*;
